refactoring

+++ src/Husky.Helpers/Helpers/LatLonHelper.cs
@@ -34,8 +34,8 @@ namespace Husky
 			var z = Math.Sqrt(latlon.Lon * latlon.Lon + latlon.Lat * latlon.Lat) + 0.00002 * Math.Sin(latlon.Lat * xPI);
 			var theta = Math.Atan2(latlon.Lat, latlon.Lon) + 0.000003 * Math.Cos(latlon.Lon * xPI);
 			return new Location {
-				Lat = (float)(z * Math.Sin(theta) + 0.006),
-				Lon = (float)(z * Math.Cos(theta) + 0.0065),
+				Lat = z * Math.Sin(theta) + 0.006,
+				Lon = z * Math.Cos(theta) + 0.0065,
 				LatLonType = LatLonType.Baidu
 			};
 		}
@@ -50,8 +50,8 @@ namespace Husky
 			var z = Math.Sqrt(lon * lon + lat * lat) - 0.00002 * Math.Sin(lat * xPI);
 			var theta = Math.Atan2(lat, lon) - 0.000003 * Math.Cos(lon * xPI);
 			return new Location {
-				Lat = (float)(z * Math.Sin(theta)),
-				Lon = (float)(z * Math.Cos(theta)),
+				Lat = z * Math.Sin(theta),
+				Lon = z * Math.Cos(theta),
 				LatLonType = LatLonType.Tencent
 			};
 		}
